+++ model/Die.c
@@ -17,7 +17,7 @@ Die * die_create(size_t numFaces) {
 }
 
 size_t die_roll(Die *d) {
-	d->last_result = arc4random_uniform(d->num_faces)+1;
+	d->last_result = (size_t) arc4random_uniform(d->num_faces) + 1;
 	return d->last_result;
 }
